+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - IcdEnvironment.GetUtcTime() to get UTC representaiton of current time.
+
 ## [9.9.0] - 2019-09-16
 ### Added
  - Added a method for converting 24 hour to 12 hour format